Move PaymentOptions to shopelectro.models
