+++ shopelectro/selenium/pages/order.py
@@ -1,7 +1,7 @@
+from shopelectro.models import PaymentOptions
 from shopelectro.selenium.elements import Input, Button
 from shopelectro.selenium.pages import Page
 
-from django.conf import settings
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 
@@ -36,11 +36,11 @@ class OrderPage(Page):
     def make_order(self):
         self.submit_button.click()
 
-    def select_payment_type(self, payment_option: settings.PaymentOptions):
-        if payment_option not in settings.PaymentOptions:
+    def select_payment_type(self, payment_option: PaymentOptions):
+        if payment_option not in PaymentOptions:
             raise ValueError(
                 'An invalid payment type provided.'
-                f'It should be one of: {settings.PaymentOptions}'
+                f'It should be one of: {PaymentOptions}'
             )
 
         item = Button(
