+++ shopelectro/settings/base.py
@@ -244,30 +244,6 @@ SITE_DOMAIN_NAME = 'www.shopelectro.ru'
 # Used to retrieve instances in ecommerce.Cart
 CART_ID = 'cart'
 
-# Used to define choices attr in definition of Order.payment_type field
-class PairIterEnum(enum.EnumMeta):
-
-    def __iter__(self):
-        for i in super().__iter__():
-            yield i.name, i.value
-
-    def __repr__(self):
-        keys = ', '.join(next(zip(*PaymentOptions)))
-        return f"<enum '{self.__name__}: {names}'>"
-
-
-class PaymentOptions(enum.Enum, metaclass=PairIterEnum):
-    cash = 'Наличные'
-    cashless = 'Безналичные и денежные переводы'
-    AC = 'Банковская карта'
-    PC = 'Яндекс.Деньги'
-    GP = 'Связной (терминал)'
-    AB = 'Альфа-Клик'
-
-    @staticmethod
-    def default():
-        return PaymentOptions.cash
-
 # It is fake-pass. Correct pass will be created on `docker-compose up` stage from `docker/.env`
 YANDEX_SHOP_PASS = os.environ.get('YANDEX_SHOP_PASS', 'so_secret_pass')
